// src/components/retro-board/types.ts
export type RetroBoardActionType = {
  SET_FROM_BOARD: 'SET_FROM_BOARD'
  ADD_CARD: 'ADD_CARD'
  UPDATE_CARD: 'UPDATE_CARD'
  DELETE_CARD: 'DELETE_CARD'
  DELETE_ALL_CARDS: 'DELETE_ALL_CARDS'
  DELETE_COMPLETED_CARDS: 'DELETE_COMPLETED_CARDS'
  TOGGLE_UPVOTE: 'TOGGLE_UPVOTE'
  MARK_DISCUSSED: 'MARK_DISCUSSED'
  ADD_ACTION_ITEM: 'ADD_ACTION_ITEM'
  TOGGLE_DONE_ACTION_ITEM: 'TOGGLE_DONE_ACTION_ITEM'
  UPDATE_ACTION_ITEM: 'UPDATE_ACTION_ITEM'
  ADD_CARD_COMMENT: 'ADD_CARD_COMMENT'
  UPDATE_CARD_COMMENT: 'UPDATE_CARD_COMMENT'
  DELETE_CARD_COMMENT: 'DELETE_CARD_COMMENT'
}